+++ src/components/DashboardEditAndDelete/DashboardEditAndDelete.jsx
@@ -1,4 +1,5 @@
 import Link from 'next/link';
+import { toast } from 'sonner';
 import { handleDeleteImgFromCloudinary } from '@/utils/handleDeleteImgFromCloudinary';
 import { handleDeleteCardFromDB } from '@/utils/handleDeleteCardFromDB';
 import { createImagesArrayForDeletingFromCloudinary } from '@/utils/createImagesArrayForDeletingFromCloudinary';
@@ -25,9 +26,12 @@ const DashboardEditAndDelete = ({ data, pathname, mutate }) => {
             <svg
                 className={styles.deleteIcon}
                 onClick={() => {
-                    const arrForDeleting = createImagesArrayForDeletingFromCloudinary(data);
-                    arrForDeleting.map(item => handleDeleteImgFromCloudinary(item));
-                    handleDeleteCardFromDB(url, mutate);
+                    if (confirm("Ви впевнені, що хочете видалити цю картку?")) {
+                        const arrForDeleting = createImagesArrayForDeletingFromCloudinary(data);
+                        arrForDeleting.map(item => handleDeleteImgFromCloudinary(item));
+                        handleDeleteCardFromDB(url, mutate);
+                        toast.success(`Картка "${data?.slug}" видалена з БД.`);
+                    }
                 }}
             >
                 <use href="/sprite.svg#icon-delete" />
